Added database for vanilla maps. Fixed small casting errors. Lobby view should now support localized scenario names.

+++ coh2-battlegrounds/coh2-battlegrounds-bin/Game/Match/Data/Events/RetreatEvent.cs
@@ -55,7 +55,7 @@ namespace Battlegrounds.Game.Match.Data.Events {
             
             // Get the veterancy change
             if (sbyte.TryParse(values[1], out sbyte change)) {
-                this.WithdrawingUnitVeterancyChange = change >= 0 ? (byte)change : 0;
+                this.WithdrawingUnitVeterancyChange = (byte)(change >= 0 ? (byte)change : 0);
             } else {
                 throw new FormatException();
             }
